requirements.txt
mesop
google-generativeai
python-dotenv
langchain>=0.2,<0.3
langchain_community>=0.2,<0.3
langchain-google-genai